finish algorithm implementation

+++ src/lib/types.rs
@@ -3,12 +3,12 @@ pub struct Problem<T> {
     // Number of servers.
     pub m: i32,
     // Finite time horizon.
-    pub t_end: usize,
+    pub t_end: i32,
     // Positive real constant resembling the switching cost.
     pub beta: f64,
     // Non-negative convex cost functions.
     // Must be defined on 1<=t<=T, 0<=x_t<=m; may return `None` otherwise.
-    pub f: fn(usize, T) -> Option<f64>,
+    pub f: fn(i32, T) -> Option<f64>,
 }
 pub type DiscreteProblem = Problem<i32>;
 pub type ContinuousProblem = Problem<f64>;
